+++ src/main/java/de/tjjf/Infrastructure/persistence/DBOperations/ImplOperations/Delete/TicketDeleteImpl.java
@@ -3,8 +3,8 @@ package de.tjjf.Infrastructure.persistence.DBOperations.ImplOperations.Delete;
 import de.tjjf.Infrastructure.persistence.DBOperations.AbstractOperations.AbstractDeleteOperation;
 import de.tjjf.Infrastructure.persistence.entities.Ticket;
 
-public class TicketDeleteImpl extends AbstractDeleteOperation<Ticket, Integer> {
-    public TicketDeleteImpl(Integer identifier){
+public class TicketDeleteImpl extends AbstractDeleteOperation<Ticket, Long> {
+    public TicketDeleteImpl(Long identifier){
         super(Ticket.class, identifier);
     }
 }
