+++ src/main/java/de/tjjf/Infrastructure/persistence/DBOperations/ImplOperations/Read/TicketReadImpl.java
@@ -3,8 +3,8 @@ package de.tjjf.Infrastructure.persistence.DBOperations.ImplOperations.Read;
 import de.tjjf.Infrastructure.persistence.DBOperations.AbstractOperations.AbstractReadOperation;
 import de.tjjf.Infrastructure.persistence.entities.Ticket;
 
-public class TicketReadImpl extends AbstractReadOperation<Ticket, Integer> {
-    public TicketReadImpl(int identifier){
+public class TicketReadImpl extends AbstractReadOperation<Ticket, Long> {
+    public TicketReadImpl(long identifier){
         super(Ticket.class, identifier);
     }
 }
