+++ src/main/java/de/tjjf/Infrastructure/persistence/DBOperations/ImplOperations/Update/TicketUpdateImpl.java
@@ -3,8 +3,8 @@ package de.tjjf.Infrastructure.persistence.DBOperations.ImplOperations.Update;
 import de.tjjf.Infrastructure.persistence.DBOperations.AbstractOperations.AbstractUpdateOperation;
 import de.tjjf.Infrastructure.persistence.entities.Ticket;
 
-public class TicketUpdateImpl extends AbstractUpdateOperation<Ticket, Integer> {
-    public TicketUpdateImpl(Ticket modelToPersist, Integer ticketId){
+public class TicketUpdateImpl extends AbstractUpdateOperation<Ticket, Long> {
+    public TicketUpdateImpl(Ticket modelToPersist, Long ticketId){
         super(modelToPersist, ticketId);
     }
 }
